pyproject.toml
[project]
name = "pro-1"
version = "0.1.0"
description = ""
authors = [
    {name = "FAIZY152",email = "fayazjani7576@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "pydantic (>=2.11.9,<3.0.0)",
    "groq (>=0.31.1,<0.32.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "src"}]

[tool.poetry.scripts]
start = "src.run:main"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
